+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fix
+- Referencing Association ignores `@segw.abap.name`
+
 ## [1.0.0] - 2025-12-06
 
 ## [1.0.0-rc.6] - 2025-12-06